fitnessbackend/requirements.txt
channels
channels-redis
daphne
djangorestframework-simplejwt
django-cors-headers
django-crontab
requests
whitenoise
django-grappelli
mysqlclient
pymysql

